Unit testing

+++ src/main/java/com/adiha/EventScheduler/services/EventsService.java
@@ -23,7 +23,6 @@ import static com.adiha.EventScheduler.utils.Constants.*;
 public class EventsService {
     private final Logger logger = LoggerFactory.getLogger(EventsService.class);
 
-
     private final EventRepository eventRepository;
     private final EventMapper eventMapper;
 
@@ -32,7 +31,7 @@ public class EventsService {
         logger.debug("Retrieving all events");
 
         if (POPULARITY.equals(sort)) {
-            return eventRepository.findAllOrderByPopularity();
+            return getEventSortedBy(order);
         }
 
         Sort sortingParameters = getSortingParameters(sort, order);
@@ -53,12 +52,10 @@ public class EventsService {
             String order) {
         logger.debug("Retrieving events with location: {} and venue: {}", location, venue);
 
-        Specification<Event> spec = Specification
-                .where(new EventByLocation(location))
-                .and(new EventByVenue(venue));
+        Specification<Event> spec = getSpecification(location, venue);
 
         if (POPULARITY.equals(sort)) {
-            return eventRepository.findAllOrderByPopularity(spec);
+            return getEventSortedBy(order);
         }
 
         Sort sortingParameters = getSortingParameters(sort, order);
@@ -72,7 +69,7 @@ public class EventsService {
         return eventRepository.save(event);
     }
 
-    public List<Event> createEvents(List<Event> events) {
+    public List<Event> createAll(List<Event> events) {
         return eventRepository.saveAll(events);
     }
 
@@ -89,7 +86,14 @@ public class EventsService {
     public void deleteEvent(String eventId) {
         logger.debug("Deleting event with id: {}", eventId);
 
-        eventRepository.deleteById(eventId);
+        if (eventId == null) {
+            throw new IllegalArgumentException("Event id cannot be null");
+        } else if (eventRepository.findById(eventId).isPresent()) {
+            eventRepository.deleteById(eventId);
+            return;
+        }
+
+        throwNotFoundException(eventId);
     }
 
     public void deleteAll(List<String> eventIds) {
@@ -100,8 +104,20 @@ public class EventsService {
         eventRepository.deleteAll(eventsToDelete);
     }
 
+    private List<Event> getEventSortedBy(String order) {
+        return DESCENDING.equals(order)
+                ? eventRepository.findAllOrderByPopularityDesc()
+                : eventRepository.findAllOrderByPopularityAsc();
+    }
+
+    private static Specification<Event> getSpecification(String location, String venue) {
+        return Specification
+                .where(new EventByLocation(location))
+                .and(new EventByVenue(venue));
+    }
+
     private Sort getSortingParameters(String sort, String order) {
-        if (sortNotAllowed(sort)) {
+        if (sort == null || sortNotAllowed(sort)) {
             throw new IllegalArgumentException("Sorting can only be done on 'creationTime' or 'startDate'");
         }
 
@@ -115,7 +131,7 @@ public class EventsService {
     }
 
     private static ResponseStatusException throwNotFoundException(String eventId) {
-        return new ResponseStatusException(
+        throw new ResponseStatusException(
                 HttpStatus.NOT_FOUND,
                 String.format("Event with uuid %s was not found", eventId));
     }
